+++ lib/polly/config.rb
@@ -11,11 +11,11 @@ module Polly
     end
 
     def self.allowed_contexts
-      rc["allowed_contexts"] || ["polly-ci", "kubernetes-admin@kubernetes"]
+      rc["allowed_contexts"] || ["polly-ci", "kubernetes-admin@kubernetes", "default"]
     end
 
     def self.image_repo
-      rc["image_repo"] || "polly-registry:443"
+      rc["image_repo"] || "polly-registry:23443"
     end
   end
 end